\name{cor_matrix}
\alias{cor_matrix}
\docType{cor_matrix}
\title{
Phylogenetic Correlation Matrix
}
\description{
Phylogenetic correlation matrix.
}
\usage{data("cor_matrix")}
\format{
A matrix.
}
\source{
See manuscript.
}
\references{
Manuscript.
}
\examples{
data(cor_matrix)
str(cor_matrix)
}
\keyword{datasets}
